backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-socketio==5.12.1
httpx==0.28.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.4
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
aiosqlite==0.20.0
